Handled division by 0 and fixed adherence to numpy dtype promotion for integer divisions

+++ tests/unit/dtype-promotion-matrix.test.ts
@@ -302,7 +302,8 @@ describe('Complete DType Promotion Matrix', () => {
       expect(boolArr.add(ones([2], 'int8')).dtype).toBe('int8');
       expect(boolArr.subtract(ones([2], 'uint16')).dtype).toBe('uint16');
       expect(boolArr.multiply(ones([2], 'float32')).dtype).toBe('float32');
-      expect(boolArr.divide(ones([2], 'int64')).dtype).toBe('int64');
+      // Division promotes to float64 (NumPy behavior)
+      expect(boolArr.divide(ones([2], 'int64')).dtype).toBe('float64');
     });
 
     it('handles operations between extreme types', () => {
